+++ src/components/modal/index.js
@@ -8,7 +8,7 @@ const Modal = (props) => {
     const [full, setFull] = useState(false);
     let className = {
         name: 'rui-modal-container',
-        hideOverplay: props.hideOverplay ? 'no-overlay' : '',
+        hideOverlay: props.hideOverlay ? 'no-overlay' : '',
         hidden: props.visible ? '' : 'hidden',
         headerReverse: props.headerReverse ? 'reverse' : '',
         centered: props.centered ? 'centered' : '',
@@ -85,7 +85,7 @@ Modal.propTypes = {
     full: PropTypes.bool,
     dark: PropTypes.bool,
     onClose: PropTypes.func,
-    hideOverplay: PropTypes.bool,
+    hideOverlay: PropTypes.bool,
     children: PropTypes.node,
     className: PropTypes.string
 }